Refactor: split routine_heartbeat.rs into focused submodules (#25) (#116)
* Refactor: split routine_heartbeat.rs into focused submodules (#25)

Split the 514-line tests/e2e_traces/routine_heartbeat.rs into five
focused submodules by concern:

- routine_cron.rs: cron-triggered routine tests
- routine_event.rs: event-triggered routine tests
- routine_system_event.rs: system event trigger tests
- routine_cooldown.rs: routine cooldown behaviour tests
- heartbeat.rs: heartbeat runner tests

Extract shared helpers into tests/support/routines.rs:
- create_test_db(), create_workspace()
- make_routine(), make_test_incoming_message()
- make_minimal_engine()
- SystemEventSpec, register_github_issue_routine()
- assert_system_event_count()

Add compile-time signature assertions in tests/support/mod.rs to
catch API drift early.

Updates tests/e2e_traces.rs to wire new submodules via #[path]
attributes following the established harness pattern.

Closes #25


* Refactor: address review feedback for routine_heartbeat split (#25)

Address code review findings:

- heartbeat.rs: Rename _db to db since the variable is used
- routine_cron.rs: Use make_minimal_engine helper instead of duplicating
  RoutineEngine construction; destructure tuple to access notify_rx
- routine_event.rs: Replace fixed 500ms sleep with polling loop that checks
  for routine completion with timeout
- support/mod.rs: Add const reference to force compile-time checking of
  routines_symbol_refs function
- support/routines.rs: Replace crate-level #![allow(dead_code)] with narrowly
  scoped #[allow(dead_code)] attributes on individual items
- support/routines.rs: Modify make_minimal_engine to return both the engine
  and the notification receiver as a tuple

All quality gates pass:
- cargo fmt -- --check: OK
- cargo clippy --tests --features libsql -- -D warnings: OK
- cargo test --test e2e_traces --features libsql: 6/6 tests pass


* Refactor: address review feedback - polling loops and error propagation (#25)

Address additional code review findings:

- routine_cooldown.rs: Replace fixed 300ms sleep with polling loop that
  queries db.list_routine_runs until the routine run appears
- routine_cron.rs: Replace fixed 500ms sleep with bounded polling loop
  that repeatedly calls db.list_routine_runs with early exit on success
- routine_system_event.rs: Replace fixed 300ms sleep with bounded polling
  loop for routine completion detection
- support/routines.rs: Change create_test_db signature to return
  Result<(Arc<dyn Database>, TempDir), Box<dyn std::error::Error>>
  and propagate errors with ? instead of .expect()
- Update all callers of create_test_db to handle the Result with .expect()

All quality gates pass:
- cargo fmt -- --check: OK
- cargo clippy --tests --features libsql -- -D warnings: OK
- cargo test --test e2e_traces --features libsql: 6/6 tests pass


* Docs: add routine test helpers documentation to testing-strategy.md (#25)

Add section 2.6 'Routine and heartbeat test helpers' to docs/testing-strategy.md
documenting the public test helper functions and structs in tests/support/routines.rs:

- create_test_db: temp libSQL database with migrations
- create_workspace: workspace backed by test database
- make_routine: routine factory with sensible defaults
- make_test_incoming_message: minimal incoming message builder
- make_minimal_engine: routine engine with notification receiver
- SystemEventSpec: system event description for tests
- register_github_issue_routine: GitHub issue routine helper
- assert_system_event_count: system event assertion helper

Include E2E test writing patterns with polling loop example for
deterministic synchronization without fixed sleeps.


---------
